Clean up app structure
