+++ lib/data/average-data.js
