+++ lib/data/calculate-coalition.js
