+++ lib/data/calculate-total.js
