+++ lib/helpers.js
@@ -47,6 +47,8 @@ function coalitionNamer(coalition) {
     return `Grand Coalition`;
   } else if(arrayEqual(coalitionPartners, ['SPD', 'LINKE', 'GRÜNE'])){
     return `Red-Red-Green`;
+  } else if(arrayEqual(coalitionPartners, ['SPD', 'FDP', 'GRÜNE'])){
+    return `Traffic Light`;
   } else if(arrayEqual(coalitionPartners, ['CDU/CSU', 'FDP'])){
     return `Black-Yellow`;
   } else if(arrayEqual(coalitionPartners, ['CDU/CSU', 'GRÜNE'])){
